Move Sphinx extension to gaphor.extensions
So it's clear it's an extension, not a core part of the
application.
