+++ gaphor/extensions/sphinx.py
